fix(orchestration): route MCP progress via factory pattern
Introduce OrchestratorFactory, a zero-state StreamingAgent that
defers Orchestrator construction to stream(). The real orchestrator
is created inside tokio::spawn where request_id flows as a direct
parameter — no Arc<RwLock> bridges needed.

This fixes MCP progress notifications being silently dropped in
orchestration mode. The inner orchestrator's MCP clients now
receive the HTTP request_id, so notifications/progress routes
to the correct SSE subscriber. Also eliminates duplicate
persistence directories and MCP connections per request.

Ref: LOG-23565
