+++ crates/aura/src/orchestration/mod.rs
@@ -18,8 +18,9 @@
 //!
 //! # Architecture
 //!
-//! The orchestrator implements `StreamingAgent`, allowing it to be used as a
-//! drop-in replacement for the standard `Agent`. It coordinates:
+//! `OrchestratorFactory` implements `StreamingAgent`, allowing it to be used as a
+//! drop-in replacement for the standard `Agent`. It creates the real `Orchestrator`
+//! lazily inside `stream()` to avoid duplicate resource allocation. It coordinates:
 //!
 //! 1. **Coordinator** - decomposes queries into plans
 //! 2. **Workers** - execute individual tasks
@@ -28,21 +29,22 @@
 //! # Example Usage
 //!
 //! ```ignore
-//! use aura::{AgentConfig, Orchestrator, StreamingAgent};
+//! use aura::{AgentConfig, OrchestratorFactory, StreamingAgent};
 //!
 //! let config = AgentConfig::from_file("config.toml")?;
 //! let agent: Box<dyn StreamingAgent> = if config.orchestration_enabled() {
-//!     Box::new(Orchestrator::new(config).await?)
+//!     Box::new(OrchestratorFactory::new(config))
 //! } else {
 //!     Box::new(Agent::new(&config).await?)
 //! };
 //!
-//! let stream = agent.stream(query, history, cancel_token).await?;
+//! let stream = agent.stream(query, history, cancel_token, "req_123").await?;
 //! ```
 
 mod config;
 mod duplicate_call_guard;
 mod events;
+mod factory;
 mod observer_wrapper;
 mod orchestrator;
 mod persistence;
@@ -58,6 +60,7 @@ pub use config::{
     ArtifactsConfig, OrchestrationConfig, TimeoutsConfig, ToolVisibility, WorkerConfig,
 };
 pub use events::{OrchestratorEvent, RoutingMode};
+pub use factory::OrchestratorFactory;
 pub use observer_wrapper::ObserverWrapper;
 pub use orchestrator::Orchestrator;
 pub use persistence::{
